LayoutTests/imported/w3c/web-platform-tests/xhr/setrequestheader-case-insensitive-expected.txt


PASS XMLHttpRequest: setRequestHeader() - headers that differ in case
FAIL XMLHttpRequest: setRequestHeader() - headers that differ in case 1 assert_equals: expected ["content-TYPE"] but got ["Content-Type"]
